+++ src/main.rs
@@ -577,7 +577,7 @@ fn create_default_scan_args() -> cli::commands::scan::ScanArgs {
         skip_aws_account_file: None,
         output_args: OutputArgs { output: None, format: ReportOutputFormat::Pretty },
         no_base64: false,
-        fast: false,
+        turbo: false,
         no_inline_ignore: false,
         no_ignore_if_contains: false,
         validation_timeout: 10,