+++ src/reporter/json_format.rs
@@ -193,7 +193,7 @@ mod tests {
             skip_aws_account: Vec::new(),
             skip_aws_account_file: None,
             no_base64: false,
-            fast: false,
+            turbo: false,
             no_inline_ignore: false,
             no_ignore_if_contains: false,
             validation_timeout: 10,