+++ src/scanner/enumerate.rs
@@ -238,7 +238,14 @@ pub fn enumerate_filesystem_inputs(
                     return Ok(());
                 }
                 progress.inc(blob.len().try_into().unwrap());
-                match processor.run(origin, blob, args.no_dedup, args.redact, args.no_base64) {
+                match processor.run(
+                    origin,
+                    blob,
+                    args.no_dedup,
+                    args.redact,
+                    args.no_base64,
+                    args.turbo,
+                ) {
                     Ok(None) => {
                         // nothing to record
                     }