+++ tests/int_github.rs
@@ -161,7 +161,7 @@ fn test_github_remote_scan() -> Result<()> {
         skip_aws_account: Vec::new(),
         skip_aws_account_file: None,
         no_base64: false,
-        fast: false,
+        turbo: false,
         extra_ignore_comments: Vec::new(),
         no_inline_ignore: false,
         no_ignore_if_contains: false,