+++ tests/int_validation_cache.rs
@@ -217,7 +217,7 @@ async fn test_validation_cache_and_depvars() -> Result<()> {
         skip_aws_account: Vec::new(),
         skip_aws_account_file: None,
         no_base64: false,
-        fast: false,
+        turbo: false,
         extra_ignore_comments: Vec::new(),
         no_inline_ignore: false,
         no_ignore_if_contains: false,